EditLabel is editable Label control.
EditLabel is editable Label control, who's text can be modified at
run-time by double-clicking on it. Under the hood, it contains a Label
control and TextBox control. It works like a Label control until you
double-click it with the mouse. When double-clicked it enters
'edit-mode' where a TextBox overlays the Label. The control will
AutoSize to the characters you type in the TextBox, updating the Label's
size at the same time. You can leave the control's edit mode by hitting
either enter, the escape key, or letting the control lose focus. After
editing, the control acts like a label again, with the updated text.